simplify hydro remix : move class HydroForRemix definition to a source .cpp file
