+++ src/solver/simulation/CMakeLists.txt
@@ -3,6 +3,7 @@ add_library(shave-peaks-by-remix-storage-gen)
 
 target_sources(shave-peaks-by-remix-storage-gen
         PRIVATE
+        hydro-for-remix.cpp
         shave-peaks-by-remix-storage-gen.cpp)
 
 target_include_directories(shave-peaks-by-remix-storage-gen
