+++ src/tests/src/solver/simulation/CMakeLists.txt
@@ -60,11 +60,9 @@ add_boost_test(test-hydro_final
 add_boost_test(tests-on-hydro-remix
         SRC
         test-hydro-remix.cpp
-        ${src_solver_simulation}/shave-peaks-by-remix-storage-gen.cpp
         LIBS
-        test_utils_unit
-        INCLUDE
-        ${src_solver_simulation}/include)
+        shave-peaks-by-remix-storage-gen
+        test_utils_unit)
 
 # ===================================
 # Tests on numspace manager
