+++ src/tests/src/solver/simulation/test-hydro-remix.cpp
@@ -2,7 +2,6 @@
 
 #define WIN32_LEAN_AND_MEAN
 
-#include <algorithm>
 #include <memory>
 #include <unit_test_utils.h>
 #include <vector>
